Refine public API compatibility snapshot
