+++ oncoref/data/therapy-benefit-toxicity-evidence.csv
@@ -1,5 +1,5 @@
 evidence_id,agent,agent_class,target_symbol,cancer_code,subtype,line_of_therapy,setting,endpoint_type,endpoint_value,benefit_tier,toxicity_tier,grade3_plus_ae_rate,discontinuation_rate,boxed_warning,major_toxicities,source_type,source_id,source_anchor,source_url,source_token,evidence_transfer,evidence_notes
-therapy_afami_cel_sarc_synovial_label,afami-cel (Tecelra),TCR-T,MAGE-A4,SARC,synovial_sarcoma,approved_biomarker_matched,unresectable or metastatic synovial sarcoma after prior chemotherapy,ORR,43.2%,high_response,high,,,CRS,cytokine release syndrome;ICANS;prolonged cytopenia;infection,label,DailyMed TECELRA label setid ab24631f-3364-46e1-8074-7244863bcbab,"TECELRA prescribing information: indication, ORR, boxed warning/adverse reactions",https://dailymed.nlm.nih.gov/dailymed/drugInfo.cfm?setid=ab24631f-3364-46e1-8074-7244863bcbab,DAILYMED_SETID:ab24631f-3364-46e1-8074-7244863bcbab,disease_matched,Afami-cel is afamitresgene autoleucel. Accelerated approval based on ORR and duration of response; requires HLA-A*02:01P/*02:02P/*02:03P/*02:06P and MAGE-A4 expression.
+therapy_afami_cel_sarc_synovial_label,afami-cel (Tecelra),TCR-T,MAGEA4,SARC,synovial_sarcoma,approved_biomarker_matched,unresectable or metastatic synovial sarcoma after prior chemotherapy,ORR,43.2%,high_response,high,,,CRS,cytokine release syndrome;ICANS;prolonged cytopenia;infection,label,DailyMed TECELRA label setid ab24631f-3364-46e1-8074-7244863bcbab,"TECELRA prescribing information: indication, ORR, boxed warning/adverse reactions",https://dailymed.nlm.nih.gov/dailymed/drugInfo.cfm?setid=ab24631f-3364-46e1-8074-7244863bcbab,DAILYMED_SETID:ab24631f-3364-46e1-8074-7244863bcbab,disease_matched,Afami-cel is afamitresgene autoleucel. Accelerated approval based on ORR and duration of response; requires HLA-A*02:01P/*02:02P/*02:03P/*02:06P and MAGE-A4 expression.
 therapy_imatinib_sarc_gist_label,imatinib,small_molecule,KIT,SARC,gist,standard_or_frontline,unresectable or metastatic KIT-positive GIST,PFS/ORR,first-line standard with durable response or disease control,major_survival,moderate,,,,edema;nausea;diarrhea;cytopenia,label,DailyMed GLEEVEC label setid 211ef2da-2868-4a77-8055-1cb2cd78e24b,GLEEVEC prescribing information: KIT-positive GIST indication/safety,https://dailymed.nlm.nih.gov/dailymed/drugInfo.cfm?setid=211ef2da-2868-4a77-8055-1cb2cd78e24b,DAILYMED_SETID:211ef2da-2868-4a77-8055-1cb2cd78e24b,disease_matched,First-line GIST standard; response and PFS vary by KIT/PDGFRA genotype.
 therapy_vorasidenib_lgg_label,vorasidenib,small_molecule,IDH1,LGG,,approved_biomarker_matched,grade 2 IDH1/2-mutant astrocytoma or oligodendroglioma after surgery,PFS,median PFS 27.7 vs 11.1 months,meaningful_pfs,moderate,22.8%,,,ALT increased;AST increased;seizure;fatigue,label,PMID:37272516; FDA VORANIGO label 218784s002lbl,INDIGO publication plus VORANIGO label safety context,https://www.accessdata.fda.gov/drugsatfda_docs/label/2024/218784s002lbl.pdf,PMID:37272516;FDA_LABEL:218784s002lbl,disease_matched,Use for IDH1/2-mutant grade 2 glioma; evidence is not a generic IDH expression benefit.
 therapy_olaparib_ov_solo1,olaparib,small_molecule,BRCA1,OV,,standard_or_frontline,newly diagnosed advanced ovarian cancer with BRCA mutation after platinum response,PFS/OS,SOLO1 sustained PFS benefit with 7-year OS follow-up,durable_rfs,moderate,,,,anemia;nausea;fatigue;myelodysplastic syndrome/acute myeloid leukemia warning,publication,PMID:36082969,SOLO1 publication,https://pubmed.ncbi.nlm.nih.gov/36082969/,PMID:36082969,disease_matched,Maintenance PARP inhibitor evidence applies to BRCA-mutated disease after platinum response.
